fix warning in ast module.
CXXFLAGS="-Wall -Wextra -Wno-unused-local-typedefs"
-Wno-unused-local-typedefs used to avoid eigen warning

+++ scilab/modules/ast/src/cpp/operations/types_or.cpp
@@ -355,7 +355,7 @@ int BoolLogicalOrBool(Bool* _pL, Bool*  _pR, Bool** _pOut)
 int SparseBoolOrSparseBool(InternalType* _pL, Bool** _pOut)
 {
     SparseBool* pL = _pL->getAs<SparseBool>();
-    if (pL->nbTrue() == pL->getSize())
+    if (pL->nbTrue() == (size_t)pL->getSize())
     {
         *_pOut = new Bool(1);
         return 0;
